+++ lib/imagekitio.rb
@@ -1,6 +1,5 @@
 require "imagekitio/railtie" if defined? Rails
 
-require 'carrierwave'
 require 'base64'
 
 require_relative './imagekitio/constant'
@@ -8,8 +7,7 @@ require_relative './imagekitio/base'
 require_relative './imagekitio/configurable'
 require_relative './imagekitio/client'
 require_relative "./imagekitio/request"
-require_relative "./imagekitio/file"
 require_relative "./imagekitio/url"
-require_relative "./imagekitio/utils/calculation"
-require_relative './imagekitio/custom_metadata_field'
-
+require_relative './imagekitio/api_service/custom_metadata_field'
+require_relative './imagekitio/api_service/file'
+require_relative './imagekitio/api_service/folder'
